+++ dflib/src/main/java/com/nhl/dflib/series/ColumnMappedSeries.java
@@ -62,6 +62,11 @@ public class ColumnMappedSeries<S, T> extends ObjectSeries<T> {
     }
 
     @Override
+    public Series<T> fillNullsFromSeries(Series<? extends T> values) {
+        return materialize().fillNullsFromSeries(values);
+    }
+
+    @Override
     public Series<T> fillNullsBackwards() {
         return materialize().fillNullsBackwards();
     }